+++ src/environments/environment.ts
@@ -1,7 +1,7 @@
 export const environment = {
   production: false,
   appVersion: '0.2.2',
-  releaseDate: '2021-11-18 23:30',
+  releaseDate: '2021-11-19 22:30',
   baseUrl: 'http://localhost:4200/',
   setAudit: false,
   firebase: {