+++ changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 
 - `feature` Add counter for favs and claps.
+- `feature` Show counter for favs and claps in event panel.
 
 ### Update